+++ pydbml/definitions/project.py
@@ -1,6 +1,6 @@
 import pyparsing as pp
 from pydbml.definitions.generic import name, string_literal
-from pydbml.definitions.common import _, note
+from pydbml.definitions.common import _, _c, n, note
 from pydbml.classes import Project, Note
 
 pp.ParserElement.setDefaultWhitespaceChars(' \t\r')
@@ -11,13 +11,13 @@ project_element = _ + (note | project_field) + _
 
 project_body = project_element[...]
 
-project = (
+project = _c + (
     pp.CaselessLiteral('project') + _ +
     name('name') + _ +
     '{' + _ +
     project_body('items') + _ +
     '}'
-)
+) + (n | pp.StringEnd())
 
 
 def parse_project(s, l, t):
@@ -30,7 +30,10 @@ def parse_project(s, l, t):
             k, v = item
             items[k] = v
     if items:
-        init_dict['items': items]
+        init_dict['items'] = items
+    if 'comment_before' in t:
+        comment = '\n'.join(c[0] for c in t['comment_before'])
+        init_dict['comment'] = comment
     return Project(**init_dict)