+++ pydbml/definitions/table.py
@@ -1,6 +1,6 @@
 import pyparsing as pp
 from pydbml.definitions.generic import name
-from pydbml.definitions.common import _, note, note_object
+from pydbml.definitions.common import _, _c, n, note, note_object
 from pydbml.definitions.column import table_column
 from pydbml.definitions.index import index
 from pydbml.classes import Table
@@ -35,7 +35,7 @@ table_settings.setParseAction(parse_table_settings)
 indexes = (
     pp.CaselessLiteral('indexes').suppress() + _ +
     pp.Suppress('{') +
-    index[1, ...] +
+    index[1, ...] + _ +
     pp.Suppress('}')
 )
 
@@ -43,15 +43,15 @@ note_element = note | note_object
 
 table_element = _ + (note_element('note') | indexes('indexes')) + _
 
-table_body = _ + table_column[1, ...]('columns') + _ + table_element[...]
+table_body = table_column[1, ...]('columns') + _ + table_element[...]
 
-table = (
+table = _c + (
     pp.CaselessLiteral("table").suppress() +
     name('name') +
     alias('alias')[0, 1] +
     table_settings('settings')[0, 1] + _ +
-    '{' + _ + table_body + _ + '}' + _
-)
+    '{' + table_body + _ + '}'
+) + (n | pp.StringEnd())
 
 
 def parse_table(s, l, t):
@@ -65,6 +65,9 @@ def parse_table(s, l, t):
     if 'note' in t:
         # will override one from settings
         init_dict['note'] = t['note']
+    if'comment_before' in t:
+        comment = '\n'.join(c[0] for c in t['comment_before'])
+        init_dict['comment'] = comment
     result = Table(**init_dict)
     for column in t['columns']:
         result.add_column(column)